db/requirements.txt
google-genai
pydantic
python-dotenv
rapidfuzz
psycopg2
sqlalchemy
fastapi